Enable streaming file output
This replaces Rubyzip with ZipTricks,
and adds `render_to` on the Document
class. `render_to` will write to any
IO object without needing it to be
rewindable or seekable.

This permits output to a socket
or to a streaming web response,
and allows memory savings when the output
is large. The deflated writer in ZipTricks
also regularly flushes the written bytes,
making sure that not too many string buffers
get accumulated.