Adds `.field` supporting :page and :numpages
This appears as helper methods for paragraphs and table cells.

Works really well in the headers and footers of subsequent commits,
but not so well in the body of a page. The information needed to
calculate the fields is apparently not ready until the pages are
completely rendered requiring a user to invoke the update fields
command manually after the doc is opened in Word or other Offiiceish
product.

We could eventually add support for the dirty flag on `w:fldChar`,
however, it does cause Word to prompt the user in an alarming fashion
when the document loads, doesn't solve the issue for documents meant
to be perfect when they are opened, and isn't as likely a use case
as headers or footers.

We should consider how best to add other fields and their formats
http://officeopenxml.com/WPfieldInstructions.php.